requirements.txt
annotated-types==0.6.0
anyio>=4.5
asgiref==3.10.0
certifi==2025.10.5
distro==1.9.0
Django==5.2.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
Markdown==3.9
openai>=2.29.0
openai-agents>=0.13.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
whitenoise==6.8.2
requests==2.32.3
gunicorn
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
beautifulsoup4>=4.12.0
googlesearch-python>=1.3.0
serpapi>=0.1.5
